[normalizer] Force declaration types to be toplevel only (part 1)
Summary:
## Motivation

The structure we use to describe inferred types (Type.t) combines cases
of *declaration* types (type aliases, class decls, modules, etc.) mixed with
*value* types (number, object, etc.). This flat representation makes it harder
to enforce invariants in the normalizer that intuitively should exist. E.g.
* Type alias, class declaration types, etc. do not nest.
* Type aliases, class declarations and modules are toplevel or parts of modules.

By the end of stack, the goal is to enforce the following structure as the output
of normalization of a type Type.t (ie. replacing Ty.t):
```
type elt =
  | Type of t
  | TypeAlias of type_alias
  | ClassDecl of symbol * type_param list option
  | InterfaceDecl of symbol * type_param list option
  | Module of symbol option * export_t
```
One of the benefit of this is that the types of expressions can be expected to be
of the sort Ty.Type. This will make it easier to map these types to our annotation
syntax, and therefore process and serialize.

## This diff

The way I approach this is by getting rid of all constructors of TypeT and ModuleT
under a recursive call to `type__` in the normalizer. This diff takes care of 2 calls to
`type_t` and `module_t`. It does so by:
1. Pulling out the logic that handles each one of these two cases in a `toplevel`
function.
2. Implementing a `Lookahead` utility that peeks at the upcoming type constructor
to determine what kind of node we're dealing with.

Reviewed By: samwgoldman

Differential Revision: D20070324

fbshipit-source-id: 70f9308df39adf452ea880920b9fa34192ab7eca