[normalizer] Force declaration types to be toplevel only (part 3)
Summary:
## Motivation

See {D20070324}

## This diff

Before, the type we inferred for the variable below
```
import * as TypeExports from './exports-type';
var M = {...TypeExports};
```
for an exports-type.js that includes a statement
```
export type T = ...;
```
would include a type-alias entry. This is awkward because `M` is otherwise
an object literal value. Accessing `T` through `M`, albeit not an error itself, yields
a useless type. This change excludes such entries from the normalized output.

Note that `T` is still part of .`TypeExports`, cause there are still legitimate uses for it.

For example:
```
var x: TypeExports.T = ...;
```
In this case `M` goes through the `toplevel` module route, which allows for such entries

Reviewed By: samwgoldman

Differential Revision: D20087387

fbshipit-source-id: b9874c13d4e7e6a9aaca0fc8f1f621bbf7894021