Fix add-comments script to add comments on JSXText
Summary:
The `add-comments` script is not currently adding comments to errors on JSXText elements, only errors on JSXElements and JSXExpressionContainers. Errors of this form can arise when a react component specifies that its children cannot be strings, but a string is passed as a child. The current `add-comments` script is failing to add comments in this case.

We can fix this by updating the `add-comments` script to add comments in an expression container above errors located on JSXText. Unfortunately this quickly exposes an edge case that we must handle. In some cases, error locs may point to the phantom JSXText after an opening JSX element tag, but before the start of any child elements, which contains just a newline and additional whitespace, like below:

```
    Loc
     |
     v
<jsx>
  JSXElement, JSXExpressionContainer, or JSXText here...
<jsx>
```

In this case we cannot safely add a comment above the error, as it may be located inside the opening element's JSX tag (if the opening tag spans multiple lines). To detect this case, we simply detect these "phantom JSXText" cases and don't write comments when the error is on a JSXText, and the error's loc starts after the last non-whitespace character of a line.

This also required fixing the logic in `getPathToLoc.js` to handle a special case of JSXText. `getPathToLoc` descends the AST, trying to find the first ast node that is completely contained within the error's loc. However this  was not accounting for the case where the error's loc only spans a portion of a JSXText (as the JSXText technically starts immediately after the opening tag of the parent JSX element and ends immediately before the closing tag of the parent JSX element, including whitespace, but the error only points to the text between the parent's opening and closing tags, not the whitespace). I relaxed this restriction when a JSXText node is encountered, and `getPathToLoc` now matches an error to a JSXText node if the error is completely contained within the node's loc. This should be safe, as JSXText nodes are leaf nodes.

Reviewed By: gkz

Differential Revision: D20072705

fbshipit-source-id: 7ead28212119e45e7bbcea05307336f079c974fc