Intern comments in update expressions
Summary:
This diff interns leading and trailing comments in update expressions, following the instructions [here](https://fb.quip.com/IgYFAbvsWtfz).

Leading comments are attached when a prefix operator is used, and trailing comments are attached when a postfix operator is used. This is because comments directly suceeding a prefix operator should be attached as leading comments to the argument, and the same is true of comments directly preceding postfix operators.

Reviewed By: pieterv

Differential Revision: D20146047

fbshipit-source-id: d2e93a2435f524db6b5e2f30ad631a50518f6dd9