Initialize as a codeActionProvider when the client supports quickfixes
Summary:
Telling the LSP client that we're a codeActionProvider means that it'll ask us for codeActions when the user hovers over errors. This'll let us provide quickfixes corresponding to type errors; specifically, the "Did you mean?" type error.

The codeActions we'll provide (both the existing "autofix exports" actions and the "did you mean?" actions being added in D20003667) will be quickfix CodeAction literals, so we initialize as a provider of quickfixes iff the client supports quickfix CodeAction literals.

Fixes #8277

Reviewed By: nmote

Differential Revision: D19919160

fbshipit-source-id: 4b204227d6425cd7de217810bbe469662f2d37d8