Simplify args to recheck_loop helper
Summary: `files_to_recheck` and `files_to_force` were never passed in to the `loop` helper, so they might as well not be arguments. `profiling` was only not passed in when the loop is first started, so IMO it's a bit clearer and harder to mess up in the future if we pass the default value in there and make it a non-optional arg.

Reviewed By: mroch

Differential Revision: D20121431

fbshipit-source-id: 9de42459f0617ba37847ed61c6d5a257929f9ed9