Test exercising dependencies ending up in the `to_check` set
Summary:
With D20097612 undone, this test causes Flow to hit that assertion and crash.

What's going on here is that if we get an IDE request for a file when Flow is in lazy mode, and that file's dependencies are not already being checked (whether focused, a dependency, or a dependent), a special kind of recheck is triggered.

Instead of passing a set of `focused` files to check, and allowing the recheck prep logic to determine which dependencies need to be checked, the set of dependencies needed is computed ahead of time, and passed explicitly as dependencies to the ordinary check codepath.

This is distinct from what happens when we focus a file explicitly, or if a set of files changes. In these cases, we pass the set of files to focus/files that have changed, and let the recheck prep logic determine the dependencies.

This only reproduces with types_first on, so it's possible that we have an existing test that would have caught this if it were running in types_first.

Reviewed By: samwgoldman

Differential Revision: D20144118

fbshipit-source-id: deb38c9bab5b7b857c472d56c7bede869b402983