[Errors] No longer group unrelated that happen to share a location
Summary:
Consider the following code:

```
//flow

var h: Array<number> = [1,2,3];
var n: Array<number> = h.concat('a', 'b', 'c'); // Error
```
Prior to this diff, checking this would produce the following error in the CLI:

```
Error -----------------------------------------------------------------------------------------

Cannot assign `h.concat(...)` to `n` because:
 - string [1] is incompatible with number [2] in array element.
 - string [3] is incompatible with number [2] in array element.
 - string [4] is incompatible with number [2] in array element.

   testfolder/test.js:4:24
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                             ^^^^^^^^^^^^^^^^^^^^^^^

References:
   testfolder/test.js:4:33
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                      ^^^ [1]
   testfolder/test.js:4:14
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                   ^^^^^^ [2]
   testfolder/test.js:4:38
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                           ^^^ [3]
   testfolder/test.js:4:43
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                                ^^^ [4]
Found 3 errors
```

While this was acceptable in a world where errors can be suppressed on any location, with the upcoming primary locations change, this becomes highly misleading because the primary location reported is not the primary location of any of the 3 errors. The error printing logic aggregated these 3 errors to find a common location, and grouped them into a single error.

When this common location is not the primary location of any of these errors, this is quite confusing, because each of the 3 above errors need to be suppressed individually, but the above appears to be a single error when it is not.

Once we include the change up the stack to print on primary locations, we report each error individually on the correct primary like this:

```
Error -----------------------------------------------------------------------------------------

Cannot assign `h.concat(...)` to `n` because string [1] is incompatible with number [2] in array element.

   testfolder/test.js:4:33
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                      ^^^ [1]

References:
   testfolder/test.js:4:14
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                   ^^^^^^ [2]

Error -----------------------------------------------------------------------------------------

Cannot assign `h.concat(...)` to `n` because string [1] is incompatible with number [2] in array element.

   testfolder/test.js:4:38
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                           ^^^ [1]

References:
   testfolder/test.js:4:14
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                   ^^^^^^ [2]

Error -----------------------------------------------------------------------------------------

Cannot assign `h.concat(...)` to `n` because string [1] is incompatible with number [2] in array element.

   testfolder/test.js:4:43
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                                                ^^^ [1]

References:
   testfolder/test.js:4:14
   4| var n: Array<number> = h.concat('a', 'b', 'c'); // Error
                   ^^^^^^ [2]

Found 3 errors
```

Reviewed By: panagosg7

Differential Revision: D20131830

fbshipit-source-id: f39ca5d37388a8d440ad04c718e80cf6d62cc05d