Add support for leading/trailing comments in `debugger`
Summary:
This diff adds internal support for comments to `Ast.Statement.Debugger`, which also involved
converting it from an atom type to a struct.

Added a new unit test, and verified that the leading and trailing comments were pulled in as expected.

NOTE: this change will require a small pattern-matching tweak in downstream dependencies

Reviewed By: mroch

Differential Revision: D20163165

fbshipit-source-id: 4d5492378b36257606971c25fdb33d21e188d35e