Intern comments in switch statements
Summary:
This diff interns leading and trailing comments around switch statements.

It seems that prettier only attaches leading comments from before the `switch` and trailing comments after the body's closing `}` to the switch statement itself, with no internal nodes (https://fburl.com/6x186txc). This diff adopts the same strategy, as other comment seems to be attached to either the discriminant or cases.

Note that there are still issues with trailing comments leading for the next statement, but this is a general problem across all comment interning at the moment.

Reviewed By: pieterv

Differential Revision: D20166116

fbshipit-source-id: 756a09f3c06975281027cb195037bb28fea799e1