Intern comments in with statements
Summary:
Interns leading and trailing comments around with statements. Does not intern any inner comments, as comments between the `with` and its body should be attached to either the with object or the body.

Note that there are still issues with trailing comments leading for the next statement, but this is a general problem across all comment interning at the moment.

Reviewed By: pieterv

Differential Revision: D20166701

fbshipit-source-id: 007453832012ac215ed27c92095d5fb99e52abf0