Optimize Trace structure
Summary:
Changes the structure of traces from
```
type step = Type.t * Type.use_t * parent * int
and t = step list
```
to
```
type step = Type.t * Type.use_t * parent
and t = step list * int
```
The int part in the trace is the only part that really needs to be kept up-to-date
even when traces are disabled (--traces 0, which is the default). This change
allows us to more aggressively optimize the Trace structure, so that
concatenating lists of (potentially empty) traces, returns an empty list, when
traces depth = 0, instead of a list of (empty traces).

Reviewed By: samwgoldman

Differential Revision: D20236440

fbshipit-source-id: 35c305d588a2827a1eb2a0092dda63df1437aab6