[lsp] scope command names per root
Summary:
LSP command names have to be unique per instance, so that when the user selects a command, VS Code knows which server to send it to (https://github.com/microsoft/vscode-languageserver-node/issues/333)

So, we append the server root path to command names when serializing, like the command `foo` becomes `foo:/home/mroch/path/to/root`.

I'm not very happy with this solution, but we can revisit the implementation after the v0.120 release. I'd like to find a way that doesn't have to thread the path everywhere.

also of note: when we receive a command, I just ignore the path and rely on the editor to not send us bogus commands. I think this is a safe assumption for now, and it avoids passing the path to even more places.

Reviewed By: vrama628

Differential Revision: D20249614

fbshipit-source-id: f5189c0d596cbbc68916af09dbb949bd376dae84