fix <PLACEHOLDER_PROJECT_URL> in tests on Macs
Summary:
on macs, the temp dir is in `/var/folders` which is a symlink to `/private/var/folders`. in our expected test output, we want to replace the project root URL with `<PLACEHOLDER_PROJECT_URL>`, but that doesn't work when it's sometimes `file:///var/folders/...` and other times `file:///private/var/folders/...`.

so, we `realpath` the temp dir before using it to run the tests.

Reviewed By: dsainati1

Differential Revision: D20283512

fbshipit-source-id: dc176fdb5f52d1a0399c3bac2cd77ca416198f2d