Intern comments in variable declaration nodes
Summary:
Interns leading and trailing comments around variable declarations.

Variable declarations can always have leading comments before the initial `var`/`let`/`const`, and also may have trailing comments after the closing semicolon. For example:

```
/* L decl */ var /* L id */ x /* T id */ = /* L num */ 1 /* T num */ ; /* T decl */
```

However variable declarations as the init of a for loop cannot have trailing comments after the semicolon, as those comments should be interpreted as leading comments of the test. Variable declarations on the left hand side of for in or for of statements also cannot have trailing comments as comments directly before the `of`/`in` will be trailing for the var decl's pattern, and comments directly after the `of`/`in` will be leading for the right hand side.

To correctly grab trailing comments when parsing I had to lift the creation of the `VariableDeclaration` out of utilities in the declaration parser, and instead move it to every call site, fetching trailing comments after the semicolon is consumed if appropriate.

Note that there are still issues with trailing comments leading for the next statement, but this is a general problem across all comment interning at the moment.

Reviewed By: pieterv

Differential Revision: D20229729

fbshipit-source-id: 6bbe97b7642c96ea3adcb020cd5505ea199a55ad