Intern comments in blocks
Summary:
Interns leading and trailing comments around blocks with the following form:

```
/* leading */ {
  const x = 1;
} /* trailing */
```

This change does not attach any inner comments, e.g. comments that trail all statements in the block:

```
{
  const x = 1;

  /* Trailing all statements */
}
```

This can be changed in the future when we settle how to handle these comments, but this diff sets up the basic structure for interning comments in blocks.

Note that there are still issues with trailing comments leading for the next statement, but this is a general problem across all comment interning at the moment.

Reviewed By: pieterv

Differential Revision: D20201572

fbshipit-source-id: 4f2d7b0d6dcc0e14231625bf150bf228d833608a