[Errors] when printing errors, the primary location should always be printed first
Summary: Previously, the error printing logic distinguished between a primary location and a "root location", which was decided based on some arcane logic that has little to do with the actual location of the error. We need to ensure that the first location printed in an error is always the primary so users know where to add suppressions, so this changes error printing to always print the primary location first, and ignore the "root loc".

Reviewed By: panagosg7

Differential Revision: D19827425

fbshipit-source-id: 45b2263f893c522fa96defd9c0028b539d7ea8f3