[Errors] Don't replace a speculation error's location with its singular branch's location
Summary:
Previously, when formatting a speculation error for printing, if we only had a single branch to the error, we'd replace the overall error's location with that of the branch. This is a bad idea. Not only does it claim that the primary location's error is positioned elsewhere than it actually is, but when checking the new `positioning.js` test file, we'd generate the following error:

```
Error ┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈ tests/error_printing/positioning.js:8:18

number literal 3 [1] is incompatible with empty [2] in type argument T of the first argument.

      8│   $fragmentRefs: $PropertyType<T, '$refType'>,
      9│ };
     10│
     11│ type X = $NonMaybeType<{|
 [1] 12│   +$refType: 3,
       :
 [2] 18│     (<T: {+$refType: empty, ...}>(?T) => ?T) &
     19│     (<TRef: 3, T: {+$refType: TRef, ...}>(T) => $FragmentRef<T>),
     20│ >;
     21│
     22│ type Props = {
     23│   selectedValue: X,
     24│ };
     25│
     26│ declare var x: $RelayProps<React$ElementConfig<(Props) => React.Node>, {}>;
```

One might expect, quite reasonably, that position `[1]` might be the primary location, or the `$PropertyType<T, '$refType'>` which is highlighted red on the CLI. In fact however, the primary location is on line 22, which only appears in the error by proximity to an included location. If the primary location were in a different file, it wouldn't appear in the error message at all. This is obviously terrible, so instead we don't replace the error's location with it's branch's, and thus correctly display the primary location.

Reviewed By: panagosg7

Differential Revision: D20108047

fbshipit-source-id: 1044008fbf28746b8af424e3c85d73cedc71228c