Add an error variant to `Saved_state_fetcher.fetch_result`
Summary:
Previously, this variant type did not distinguish between the case where we tried to find saved state and failed (e.g. local fetcher) and the case where we simply did not bother to look for saved state (e.g. the dummy fetcher).

This leads to an inconsistency where a failure to find the saved state and a failure to read the saved state are handled in completely different ways. This is illustrated by the test added in D20252511.

This diff adds the additional variant needed to distinguish between failing to find saved state and choosing not to look for saved state. For now, they are handled identically, but future diffs will change that.

Reviewed By: mroch

Differential Revision: D20236637

fbshipit-source-id: 5da544d66c3982af95ea568a09754d615a640ff0