[Errors] only allow suppressions on an error's primary location
Summary:
Instead of iterating over all of the locations in a printable error, we take the head of the list (the primary location) and only allow suppressions on this location.

One thing that may be slightly surprising about this change is that suppressing an object-like type when there is an error on one of its fields no longer suppresses that error, because the primary location of the error is on the field, rather than the object as a whole. Consider the following:

```
const React = require('react');

type Props = {|
    x : number
|};
declare function MyComponent(props: Props): React.Node;

// $FlowFixMe unused
<MyComponent
x={true} // error is here
/>
```

This FixMe does not suppress the error because the primary location is on the line `x={true}`, not on the JSX expression.

Reviewed By: panagosg7

Differential Revision: D19780657

fbshipit-source-id: 9bdbf1d9f9e7585a5578195a41242a3f87b54c26