Attach comments around parenthesized expression to inner expression
Summary:
When a parenthesized expression has leading or trailing comments, we should attach those leading and trailing comments to the expression node inside the parentheses. This will simplify comment attaching, as it will let us guarantee that all leading/trailing comments of expressions will be attached to that expression's AST node. Without this, every node that has an expression as a child must have additional handling for whether the expression child is parenthesized or not.

For example, consider a binary operator where both the left and right hand side are parenthesized. This is the required handling if leading and trailing comments are not attached to the left and right hand expression nodes.

```
/* Leading binary */ (1) /* Inner? binary */ + /* Inner? binary */ (2) /* Trailing binary */
```

Notice that we need additional inner comments in the binary expression node in case parentheses are used. This will be true for all nodes that contain an expression child. Additionally the comments should likely actually be attached to the left and right expressions themselves.

Instead, the following is the comment attachment with this change:
```
/* Leading literal */ (1) /* Trailing literal */ + /* Leading literal */ (2) /* Trailing literal */
```

The only way to attach comments to the entire binary expression would be to use parentheses, which seems sensible:

```
/* Leading binary */ (1 + 2) /* Trailing binary */
```

This required adding an `add_comments` utility to the expression parser which adds additional comments to an expression  node if it is wrapped in parentheses.

Reviewed By: mroch

Differential Revision: D20236723

fbshipit-source-id: 664c32e99da69fabba9d14794a5dadd05251430c