Intern comments in template nodes
Summary:
Interns leading and trailing comments around template literals and tagged templates.

Template literals can have leading and trailing comments of the following form:

```
/* L template */ `template` /* T template */
```

Tagged template literals can also have leading and trailing comments, but only if parentheses are used. Otherwise the leading comments will be attached to the inner expression (tag) node, and the trailing comments will be attached to the inner template literal node:

```
/* L tagged template */ (tag`template`) /* T tagged template */
/* L tag expr */ tag`template` /* T template */
```

Note that there are still issues with trailing comments leading for the next statement, but this is a general problem across all comment interning at the moment.

Reviewed By: mroch

Differential Revision: D20262299

fbshipit-source-id: e4595b886e8caf1d83de8951f2817e48a3d98186