Intern comments in meta property nodes
Summary:
Interns leading and trailing comments in meta property nodes.

Since meta property nodes begin and end with identifiers, leading comments will usually be attached to the first identifier and trailing comments will usually be attached to the trailing identifier:

```
/* L new */ new /* T new */ . /* L target */ target /* T target */
```

However if wrapped in parentheses, the leading and trailing comments will be attached to the meta property node itself:

```
/* L meta property */ (new.target) /* T meta property */
```

This diff also adds handling for meta properties in the ast mapper and ast differ.

Reviewed By: mroch

Differential Revision: D20292477

fbshipit-source-id: 67c00d853a8ce0f02749895af00f16a85df31381