add type definition for Promise.allSettled()
Summary:
There is a newer API, `Promise.allSettled()`, that is similar to `Promise.all()`,
though it gives the caller more flexibility in dealing with rejected promises:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/allSettled

I would like to use this in my own code, so I am adding a Flow definition.
I modeled this after the type definition for `Promise.all()`, which seemed to
work nicely.

Reviewed By: nmote

Differential Revision: D20342965

fbshipit-source-id: 7d91cb412d831322a5d6e4b27ba78e3a84c00378