Intern comments in JSX expression containers
Summary:
Intern leading and trailing comments around JSX expression containers.

Leading and trailing comments can appear adjacent to JSX expression containers in JSX attributes. Note that the comments inside the expression container will be attached to the inner expression:

```
<div attribute=/* L JSX expr */ { /* L expr */ expr /* T expr */ } /* T JSX expr */ />
```

JSX expression containers that appears as children of a JSX element cannot have leading or trailing comments.

Note that trailing comments of JSX expression containers in JSX attributes could also be attached to the next JSX attribute name. This will be fixed in the future once we have an API for eating portions of comments, as this same problem appears for the trailing comments on statements.

Reviewed By: pieterv

Differential Revision: D20316862

fbshipit-source-id: 96e849a4bd6a9ef8438c431d346d8bfe40d64219