Intern comments in JSX spread children
Summary:
Interns comments in JSX spread children. JSX spread children appear as JSX children, so that cannot have comments directly preceding or suceeding them. However they can have comments between the opening brace and the ellipsis. We can attach these as the "leading" comments for JSX spread children. The comments directly preceding and succeeding the expression that is spread will be attached to the expression itself:

```
<div>{/* L spread */... /* L expr */ expr /* T expr */}</div>
```

Note that this required creating a new record type in the ast for JSX spread children so that we can store both the expression and comments, as no record type previously existed.

Reviewed By: pieterv

Differential Revision: D20317867

fbshipit-source-id: 567880ce9695c52f26bf4c4eb013f92c4abb5fa1