Intern comments in JSX spread attributes
Summary:
Interns comments around JSX spread attributes. Comments that directly precede and succeed the braces of the JSX spread attribute will be attached as trailing and leading comments, respectively.

```
<div attribute=/* L JSX spread */ {.../* L expr */ expr /* T expr */} /* T JSX spread */ />

```

Note that this does not attach comments that appear between the opening brace the the ellipsis. If we want to support these, they can be added later as internal comments.

Also note that there is still ambiguity about whether leading and trailing comments for JSX attributes should be attached to the previous or next node. This will be fixed in the future once we have an API for eating portions of comments, as this same problem appears for the trailing comments on statements.

Reviewed By: pieterv

Differential Revision: D20319152

fbshipit-source-id: 75c64b0989f1a2aeadbd21478d916abdf73dd091