Intern comments in private name nodes
Summary:
Interns leading comments in private name nodes. Private names begin with a `#` and end with an identifier, so comments directly preceding the `#` will be attached as leading comments. It is a syntax error to have a comment between the `#` and the identifier, but comments succeeding the identifier will be attached as trailing comments for the inner identifier:

```
/* L private name */ #ident /* T id */
```

This adds support for comment attachment for private names in member expressions and private class fields.

Note that I created a new record type for `PrivateNames` that can store both the inner identifier and the attached comments.

Reviewed By: mroch

Differential Revision: D20355492

fbshipit-source-id: 1499df6adb0bab43e07eacb0e54fc6ad91607740