Fix deadlock when running an external process
Summary:
Previously, we would wait for the external process to exit before beginning to read its output. This means that if the external process writes enough to fill up the OS buffer, its `write` calls begin to block. Because we are waiting for it to exit, and it is waiting for us to read, we reach a deadlock.

This fixes the issue by creating all of these promises in parallel. This way, we start the reads at the same time as we start watching for the process to exit.

Reviewed By: samwgoldman

Differential Revision: D20424776

fbshipit-source-id: e1fc62ff36bfdf72bdda064cda01215b20158384