Fix types-first signatures for CJS requires introduced with var
Summary:
Before this diff, Flow in types-first mode would allow programs like this:

```
var C = require("C");
// [snip]
module.exports = (new C(): C);
```

In the above, since `C` is bound using `var`, the snipped code might re-assign
the bound value in a way that the types-first signature does not capture.  This
behavior is unique to top-level requires such as this. All other variable
bindings must be either `const` or annotated.

This unique behavior arises due to the way we copy in requires from the file
sig. The signature builder skips over import/export statements during its own
pass over the AST, but then copies that information from the file sig, which
gathers that information during its own pass.

For ES import/export statements, this doesn't cause any problem. The signature
builder is able to completely skip over them, because they are easily identified
by the Ast.Statement.t variant.

For CJS requires, however, the signature builder actually does create entries
for the variable declaration as if it were any other variable declaration, but
then those entries are overwritten.

It turns out that the original entries are entirely sufficient to capture
the requires, since there is already logic to model "dynamic requires" which
might appear in any inline expression.

Using the dynamic requires logic does change the extracted signatures, because
dynamic requires are "outlined" and lifted to annotations using `typeof`. In
practice, this does not seem to have any effect on the checking behavior. This
change in output can be seen in the diff in expected test output.

Reviewed By: panagosg7

Differential Revision: D19985936

fbshipit-source-id: 3f351fb0e84bc3edb664f8ee0c913e12da68e3d4