[enums] Adjust position of invalid enum member access reason
Summary:
Before, we were pointing at the entire expression, e.g. `E.A`, now we point at just the member itself, e.g. the `A` in `E.A` - this mirrors how these things are reported for invalid property accesses in objects.

In addition, we rename the main reason in the error to just `reason`, following the pattern set in other errors.

In addition to adding consistency with missing property errors, this also enables us to more easily implement quick fix suggestions for incorrect member accesses.

Reviewed By: dsainati1

Differential Revision: D20371181

fbshipit-source-id: 96d4c94e8db6304603c60832395b6a8ddf7fde48