Intern comments in type aliases and opaque type aliases
Summary:
Interns leading and trailing comments in type aliases and opaque type aliases. Comments directly preceding the initial `type` or `opaque` keywords will be attached as leading comments, and comments directly succeeding the closing semicolon will be attached as trailing comments.

```
/* L alias */ type Foo = boolean; /* T alias */
```

Also note that comments between the `opaque` and `type` will be attached as leading comments:

```
/* L opaque */ opaque /* L opaque */ type Foo = boolean;
```

Comments around the type alias' name will be attached as comments on that identifier (or its type params). Similarly, comments around the type on the right side will be attached as comments on that type:

```
/* L alias */ type /* L id */ Foo /* T id */ = /* L type */ boolean /* T type */ ; /* T alias */
```

Note that there are still problems with trailing comments being attached to the next statement as well, but this is a general problem across all statements.

Reviewed By: mroch

Differential Revision: D20372490

fbshipit-source-id: 22bd4064310a4b2e36d45f8e160e18d616b6731d