Intern comments in object types
Summary:
Interns leading and trailing comments on object type nodes. Comments that directly precede and succeed the braces will be attached as leading and trailing comments, respectively. Comments within the braces will be attached to the properties:

```
type Foo = /* L obj */ {} /* T obj */;
```

Note that this does not handle comments inside an empty object, as these are internal comments which have not yet been implemented.

Reviewed By: mroch

Differential Revision: D20377492

fbshipit-source-id: 24fd0937abf85a22051b1d9d18811efd038a5f33