Intern comments in tuple type nodes
Summary:
Interns leading and trailing comments around tuple type nodes. Comments that directly precede the opening bracket will be attached as leading comments, and comments that directly succeed the closing bracket will be attached as trailing comments. Comments between the brackets will be attached to the inner list of types:

```
/* L tuple */ [ /* L string */ string /* T string */ ] /* T tuple */
```

Note that this required creating a new record type in the AST for tuple type nodes so that we can keep track of both the inner list of types and the attached comments.

Reviewed By: gkz

Differential Revision: D20391455

fbshipit-source-id: 1e5cca8a566ecdcd6efceb33ec13d8b287a156bc