Intern comments in typeof nodes
Summary:
Interns leading comments for typeof types. Comments that directly precede the `typeof` will be attached as leading comments. Comments between the `typeof` and type will be attached as leading comments for the type, and comments succeeding the type will be attached as trailing comments for the type:

```
/* L typeof */ ?=typeof /* L type */ string /* T type */
```

Note that this required creating a new record type in the AST for typeof type node so that we can keep track of both the type argument and the attached comments.

Reviewed By: mroch

Differential Revision: D20390675

fbshipit-source-id: d8cca4b32417b66e7ce60c6ac617135f57f9ef00